requirements.txt
tensorflow-gpu==1.7
numpy
scipy
opencv-python
tqdm
pyyaml
flake8
h5py
jupyter
coloredlogs
pandas
